Refactor project

+++ bl/src/main/java/com/techwizzle/site/repositories/ContactFormMessageRepository.java
@@ -3,6 +3,6 @@ package com.techwizzle.site.repositories;
 import com.techwizzle.site.model.ContactFormMessage;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ContactFormMessageRepository extends CrudRepository<ContactFormMessage, Long> {
+public interface ContactFormMessageRepository extends CrudRepository<ContactFormMessage, Integer> {
 
 }
