+++ bl/src/main/java/com/techwizzle/site/repositories/GameRepository.java
@@ -3,6 +3,6 @@ package com.techwizzle.site.repositories;
 import com.techwizzle.site.model.Game;
 import org.springframework.data.repository.CrudRepository;
 
-public interface GameRepository extends CrudRepository<Game, Long> {
+public interface GameRepository extends CrudRepository<Game, Integer> {
 
 }
